file 1, print node length of doubly linked list, tested and compiled, test runs larger number, cheker did not like that, used int instead of int, still getting a larger number

+++ doubly_linked_lists/1-dlistint_len.c
@@ -9,7 +9,7 @@
 **/
 size_t dlistint_len(const dlistint_t *h)/**print length of LL**/
 {
-	size_t length;/**use length for readability**/
+	int length;/**use length for readability**/
 
 	while (h)/**while DLL--**/
 {
